requirements.txt
# Web 框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# 数据库
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0

# HTTP 客户端
httpx>=0.25.0

# 数据验证
pydantic>=2.0.0

# 环境变量
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
